Rename to `mouse_over`
